Remove application metadata
This project was discontinued upstream and forks which
continued it never updated that info or claimed to be
actual forks of the initial project.

The original author requested his name and contact
details to be removed also, so that present and future
users would stop contacting him about this software.